move input validation to pipe abstract
