+++ pipelex/pipe_operators/llm/pipe_llm_blueprint.py
@@ -6,7 +6,7 @@ from pipelex.cogt.llm.llm_setting import LLMModelChoice
 from pipelex.cogt.templating.template_category import TemplateCategory
 from pipelex.cogt.templating.template_preprocessor import preprocess_template
 from pipelex.core.pipes.pipe_blueprint import PipeBlueprint
-from pipelex.pipe_operators.llm.input_validation_helpers import is_input_used_by_variables, is_variable_satisfied_by_inputs
+from pipelex.core.pipes.validation import is_input_used_by_variables, is_variable_satisfied_by_inputs
 from pipelex.tools.jinja2.jinja2_errors import Jinja2DetectVariablesError
 from pipelex.tools.jinja2.jinja2_required_variables import detect_jinja2_required_variables
 from pipelex.types import StrEnum
